Theoretical (EN)
*Source data:*

There is a form with a field and a button. A date is entered in the field. When the button is clicked, it calculates whether the entered year is a leap year or not. The result is displayed in the interface.

*Task:*

Using test analysis and test design techniques, create test cases.

*Additional questions:*

1. What test design and test analysis techniques exist?
2. What is the difference between test design and test analysis?
3. What are the pros and cons of test cases?
4. How to determine if a feature is tested?