WEB
There is a resource where the file **[http://test.com/test.exe](https://h.yandex-team.ru/?http%3A%2F%2Ftest.com%2Ftest.exe)** is located (this is a hypothetical resource).

On 3 machines, this file is simultaneously downloaded through the browser. The browser on all machines is Google Chrome, and the operating systems are of the same version (e.g., Win7).

As a result of the download (in the Downloads folder), the test.exe file has version v2 on 2 machines, and version v1 on one machine.

*Task:*

Explain why this could have happened. How will you investigate the problem (using which tools)?

*Additional questions:*

1. What does an HTTP request consist of?
2. What methods are there?
3. How to choose which browsers to test the project in?